fix: 파일 첨부 에러 해결, 파일 미리보기 형식으로 변경

+++ drive/drive_init.py
@@ -18,11 +18,15 @@ def init_drive_api():
 # 파일 업로드 함수
 def upload_to_drive(file, file_name, folder_id):
     drive_service = init_drive_api()
-    # 임시 파일에 업로드된 파일을 저장
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        temp_file.write(file.read())
-        temp_file_path = temp_file.name
+     # 임시 파일에 업로드된 파일을 저장
+    temp_file_path = None
     try:
+        # 임시 파일 생성 및 저장
+        with tempfile.NamedTemporaryFile(delete=False) as temp_file:
+            temp_file.write(file.read())
+            temp_file_path = temp_file.name
+
+        # MediaFileUpload로 업로드
         media = MediaFileUpload(temp_file_path, mimetype=file.content_type, resumable=True)
         file_metadata = {
             "name": file_name,
@@ -31,12 +35,28 @@ def upload_to_drive(file, file_name, folder_id):
         uploaded_file = drive_service.files().create(
             body=file_metadata,
             media_body=media,
-            fields="id, webContentLink"
+            fields="id, webViewLink"
         ).execute()
+
+        # 파일 권한 설정
+        permission = {
+            "type": "anyone",
+            "role": "reader"
+        }
+        drive_service.permissions().create(
+            fileId=uploaded_file.get("id"),
+            body=permission
+        ).execute()
+        
         return {
             "id": uploaded_file.get("id"),
-            "webContentLink": uploaded_file.get("webContentLink")
+            "webViewLink": uploaded_file.get("webViewLink")
         }
+
     finally:
         # 임시 파일 삭제
-        os.remove(temp_file_path)
+        if temp_file_path and os.path.exists(temp_file_path):
+            try:
+                os.remove(temp_file_path)
+            except PermissionError:
+                print(f"파일 삭제 실패: {temp_file_path}. 다른 프로세스에서 사용 중일 수 있습니다.")
